- lib64 fix for old-style link system for iso live systems
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@1212 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ modules/KIWIImage.pm
@@ -928,8 +928,11 @@ sub createImageLiveCD {
 				$this -> restoreCDRootData();
 				return undef;
 			}
-			my @rodirs = qw (bin boot lib opt sbin usr);
+			my @rodirs = qw (bin boot lib lib64 opt sbin usr);
 			foreach my $dir (@rodirs) {
+				if (! -d $imageTree/$dir) {
+					next;
+				}
 				$data = qxx ("mv $imageTree/$dir $imageTreeReadOnly 2>&1");
 				$code = $? >> 8;
 				if ($code != 0) {
@@ -1080,8 +1083,11 @@ sub createImageLiveCD {
 			$this -> restoreCDRootData();
 			return undef;
 		}
-		my @rodirs = qw (bin boot lib opt sbin usr);
+		my @rodirs = qw (bin boot lib lib64 opt sbin usr);
 		foreach my $dir (@rodirs) {
+			if (! -d $imageTree/$dir) {
+				next;
+			}
 			$data = qxx ("cp -a $imageTree/$dir $imageTreeReadOnly 2>&1");
 			$code = $? >> 8;
 			if ($code != 0) {
@@ -3054,8 +3060,11 @@ sub restoreSplitExtend {
 		return $imageTreeReadOnly;
 	}
 	$kiwi -> info ("Restoring physical extend...");
-	my @rodirs = qw (bin boot lib opt sbin usr);
+	my @rodirs = qw (bin boot lib lib64 opt sbin usr);
 	foreach my $dir (@rodirs) {
+		if (! -d $imageTreeReadOnly/$dir) {
+			next;
+		}
 		my $data = qxx ("mv $imageTreeReadOnly/$dir $imageTree 2>&1");
 		my $code = $? >> 8;
 		if ($code != 0) {
